Allow passing a config override file through the configuration endpoint. (#1452)
## Why make this change?
We want to be able to provide the configuration file and a set of
overrides when configuring the engine at runtime. This will allow us to
override any existing or new setting without requiring an other update
to the configuration endpoint.

closes #1484

## What is this change?
- This adds a second configuration endpoint at /configuration/v2. This
endpoint expects a "Configuration Overrides" which matches the schema of
the configuration file.
- Update tests to validate both the old and the v2 endpoint.

## How was this tested?
- [x] Integration Tests
- [x] Unit Tests

## Remaining work
- Once SWA integration is updated to use the new configuration endpoint,
the old one will be removed.