Fix merge config file not available issue (#1493)
## What is the issue?
- This issue occurs specifically when the environment variable is set
and `environmentBasedConfigFile` is not present and the user tries to do
`dab start`. It will search for the merged file and fail directly if the
merged file is not present. Ideally it should not and rather look for
the default file.
- Even though the merging was not happening and the method
`TryMergeConfigsIfAvailable` was returning false, the out file was not
empty and it was containing the name of expected mergeConfigFile.
 
## Why make this change?
- the merge config file name is generated and set as the config to be
used even if the merged config file is not available.
- This will make sure merge config file name is not used if it is not
available

## What is this change?

- Moving the line which creates the merged file name inside the
validation check for existence of the base and environment config.

## How was this tested?

- [x] Unit Tests