Add support for adding base-route for Runtime (#1506)
## Why make this change?
SWA or any other front end can interact with DAB using a path-based
routing mechanism, where they may use a path (`base-route`) to redirect
requests to DAB (refer to the bug encountered:
https://github.com/Azure/data-api-builder/issues/1420). This
`base-route` gets stripped off when the request lands at DAB.
Consequently, DAB is not even aware of anything like a `base-route`
being used by the front end.
This poses a problem in generation of the `nextLink`, where the
`nextLink` which DAB generates, does not contain the `base-route` and
consequently when the front end receives the next request, it does not
find the base route in the request URL and hence, has no idea where to
redirect the request to.

Using the newly added option to configure base route for rest requests
via the config (CLI option - `--runtime.base-route`), the front-end like
SWA can make DAB aware that they are using a base route to redirect
requests to DAB. DAB can then accordingly account for the base route
while generating the `nextLink` for paginated responses.
## What is this change?

1. Added feature to specify `base-route` in global runtime settings in
the config which can be via CLI as well using `--runtime.base-route
`option in the `dab init` command. By default, it will be initialized as
null. When the base-route is a non-empty string, it will be included in
the generated nextLink. This feature is only supported when
`authentication provider == StaticWebApps`. We throw an exception when
base-route is configured for non-SWA providers.
2. Extended functionality of `Utils.IsApiPathValid(string? apiPath,
ApiType apiType)` method to `Utils.IsUriComponentValid(string?
uriComponent)` because similar validations are required for api
properties: `rest.path`/ `runtime.base-route`/ `graphql.path`, all of
which are part of the request URI.
3. Renamed `RuntimeConfigValidator.ValidateRestPathForRelationalDbs()`
to `RuntimeConfigValidator.ValidateRestURI()`.
4. Renamed `RuntimeConfigValidator.ValidateGraphQLPath()` to
`RuntimeConfigValidator.ValidateGraphQLURI()` so that we follow same
convention between rest and gql.

## Q&A

1. Will this require updating an already created configuration file? 
No, the existing config file will work perfectly fine. The base-route is
totally an optional feature.
2. Will a default value be implicitly used if the new base path config
property is not set?
The default value for the base-route is null. This essentially means
that base-route will not come into the picture.
3. What is the default value and does the value differ between dev/prod
swa/non-swa?
The default value will be same irrespective of dev/prod/swa/non-swa
environment and equal to null.

So to say, the **change is not a breaking change and we are totally
backwards compatible**.

## How was this tested?
- [x] Integration Tests - Was added to `ConfigurationTests.cs` to assert
that the generated nextLink contains the base-route.
- [x] Unit tests - were added to `ConfigValidationUnitTests.cs` to
confirm that we throw proper exceptions when runtime base-route is
formatted incorrectly (contains reserved characters or doesn't start
with a '/') or when base-route is configured for non-SWA authentication
providers.

---------

Co-authored-by: Aaron Powell <me@aaron-powell.com>
Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>