[BugFix] Handle configs which have missing options for MsSql (#1580)
## Why make this change?

- With the recent refactor #1402, if `options` property is not specified
in the config, after deserialization the `Options` dictionary of
`DataSource` object is `null`
- This leads to a null dereference when trying to
`GetTypedOptions(MsSqlOptions)`
[here](https://github.com/Azure/data-api-builder/blob/bda0d6551f8f75fc7b98f888bdfce1f3dd317378/src/Core/Resolvers/MsSqlQueryExecutor.cs#L76C70-L76C70).

## What is this change?

- Make `Options` a nullable property since its not required as per the
schema.
- Handle all the instances of Options considering it could be null.

## How was this tested?

- Manually running dab.
- Automated tests:
1. Removed the `options: set-session-context` from the initial config
used in the CLI tests to verify deserialization results in a null
`Options` dictionary as demonstrated by the Snapshots
2. Modified tests which create a custom config and start the engine by
setting `Options` to null and verifying those continue to pass.
Note: The checked in configuration file for MsSql sets the session
context to true for integration tests of row level security feature. So,
did not modify that file even though options is optional for MsSql.