Add order by and filter tests for missing SQL Types to GQL (#1557)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/1533
Our test coverage was missing order by and filter tests for a most of
the SQL Types we support in GQL. We add those tests. Adding `DateTime`
support for PostgreSql will take additional work. Those tests are
bypassed for now.

## What is this change?

Adds testing for `order by` and `filter` for all the DB types within
GQL, with the exception of `DateTime` for `PostgreSql`.

## How was this tested?

Ran against our test suite.

## Sample Request(s)

The valid GQL is included within the test cases. These can be copied if
you wish to run manually.