OpenAPI document resolves custom configured REST path (#1653)
## Why make this change?

- Closes #1633
- The wrong 'path' value is generated in the OpenAPI description
document created. This was observed when looking at the `/swagger`
endpoint to see the REST endpoint documentation.
- When an entity's REST path property is explicitly set AND includes a
starting slash `/`, the openapidocumentor does not properly create the
PATH used to access the entity in the REST endpoint.

## What is this change?

- Correctly process an entity's REST path configuration. Before this
change, the method at fault was improperly calling `Substring()` on the
wrong variable. Although the variable was previously `restPath`, the
value at the time of calling substring was the top-level entity name
which did not have a starting slash `/`.
- The variable `restPath` is modified depending on whether the entity
has explicit rest path config. If no explicit rest path is configured,
the top level entity name is returned as the resolved rest path.
- Now, calling `entityRestSettings.Path.TrimStart('/')` will correctly
remove a starting slash IF one exists. This accommodates the following
formatting of the explicit rest path:

1. `/customRestPath`
2. `customRestPath`
3. `//customRestPath` -> while this option is not advisable to use in
practice, we still support it and this PR is not meant to change that
behavior. But for complete test coverage, this case is tested.

## How was this tested?

- [x] Integration Tests