Fix Config File Path  Issue (#1681)
## Why make this change?

- Closes #1674 ConfigFile was not found when running in the Azure
Container Apps.

## What is this change?

- Because of the way we used to fetch the cofigFileName, the path was
lost, due to which using config file present in a different directory
was giving `file not found` error.
- So we updated the method to fetch the configFileNamewihoutExtension to
keep the original path.
- To run Azure Container Apps `--ConfigFileName` is provided with the
full path name to the mounted config file.

NOTE:
* This issue might be caught in Az Container apps but this is not
specifically related to Containers, because the current file name
generation process was itself incorrect in 0.8.* release.
* When we try to get the `fileNameWithoutExtension` in the method
`GetFileNameForEnvironment`, we used to lose directory information, for
example: if baseConfigFile is`dab/configs/dab-config.json`, then
`configFileNameWithoutExtension` was coming as `dab-config`, which is
incorrect.
* This issue started happening from 0.8.* because in the earlier release
(0.7.6), if config file name is provided by user, we directly used it
without searching for overrides and environments, unlike 0.8.*. And the
precedence check used to happen for cases where config file is not
provided by user.
* Now, since 0.8.*, we started using the method
`GetFileNameWithoutExtension`, which is incorrect causing File not found
issue whenever the config file is not in current directory.
* The method `DoesFileExistInCurrentDirectory()` checks if the file
exists by combining the current directory with the fileName. Now if file
is existing in some other directory inside current directory, then for
proper searching the file should contain complete info about the
directory. So, if current dicrectory is `C:/dab`, and file is in
`C:/dab/configs/dab-config.json`, then the file name should either be
full path or `/configs/dab-configs.json` for it to be searched.

## Additional Change
1. Fixing name of variables and function name to correctly reflect their
behavior.
2. Fixing issue with not honoring the user provided config file, (caused
after 0.8 release). (Putting in this PR, because this PR aims at fixing
how we are using the config files and not just a bug fix.)

## How was this tested?

- [X] Running In Azure Container Apps
BEFORE:

![image](https://github.com/Azure/data-api-builder/assets/102276754/183d9399-9822-4435-8322-e7bfc5016e1e)

AFTER:

![image](https://github.com/Azure/data-api-builder/assets/102276754/fb7010c9-4410-48cf-9146-9cc757f209a0)

- [X] Local Testing
BEFORE:

![image](https://github.com/Azure/data-api-builder/assets/102276754/742cfd03-ee54-4f0c-8ea4-9e1d2303bbf8)

AFTER:

![image](https://github.com/Azure/data-api-builder/assets/102276754/bef4d153-9f5a-470b-8fbb-27ed7219fac6)

- [X] Unit Tests

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>